src/token-manager/requirements.txt
azure-core==1.28.0
azure-functions==1.15.0
azure-identity==1.13.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.6
cryptography==41.0.2
Flask==2.3.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.7.0
MarkupSafe==2.1.3
msal==1.22.0
msal-extensions==1.0.0
portalocker==2.7.0
pycparser==2.21
PyJWT==2.7.0
pywin32==306; sys_platform == 'Windows'
requests==2.31.0
six==1.16.0
typing_extensions==4.7.1
urllib3==2.0.3
Werkzeug==2.3.6
win32-setctime==1.1.0
